Create vent view component
